# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pi Zero 2W Medicine Tracker API
  description: |
    RESTful API for medicine tracking and management on Pi Zero 2W.
    This API provides endpoints for managing medicines, tracking adherence,
    and configuring application settings.
  version: 1.0.0
  contact:
    name: Pi Zero 2W Project
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: http://{hostname}:5000/api/v1
    description: Pi Zero 2W device server
    variables:
      hostname:
        default: pizero.local
        description: Hostname or IP address of the Pi Zero 2W device
...
paths:
...
  /medicines:
    get:
      tags:
        - Medicines
      summary: List Medicines
      description: Get list of medicines with optional filtering and pagination
      operationId: listMedicines
      parameters:
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
            default: true
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: Items per page (max 100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          description: Field to sort by
          schema:
            type: string
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated list of medicines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMedicineList'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Medicines
      summary: Create Medicine
      description: Create a new medicine
      operationId: createMedicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineInput'
            examples:
              basic:
                summary: Basic medicine
                value:
                  name: Aspirin
                  dosage: 500mg
                  frequency: Twice daily
                  stock: 100
              advanced:
                summary: Advanced medicine with time windows
                value:
                  name: Metformin
                  dosage: 1000mg
                  frequency: Daily
                  time_windows:
                    - hour: 8
                      minute: 0
                    - hour: 20
                      minute: 0
                  stock: 60
                  low_stock_threshold: 20
      responses:
        '201':
          description: Medicine created successfully
          headers:
            Location:
              schema:
                type: string
              description: URL of created resource
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Medicine'
                  message:
                    type: string
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate medicine ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Medicine:
      type: object
      properties:
        id:
          type: string
          description: Unique medicine identifier
        name:
          type: string
          description: Medicine name
        dosage:
          type: string
          description: Medicine dosage
        frequency:
          type: string
          description: How often to take
        stock:
          type: integer
          description: Number of pills in stock
        low_stock_threshold:
          type: integer
          description: Alert when stock falls below this value
        time_windows:
          type: array
          items:
            type: object
            properties:
              hour:
                type: integer
              minute:
                type: integer
          description: Scheduled times for taking medicine
        active:
          type: boolean
          description: Whether medicine is currently active
        created_at:
          type: string
          format: date-time
          description: When medicine was added
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    MedicineInput:
      type: object
      required:
        - name
        - dosage
        - frequency
        - stock
      properties:
        name:
          type: string
          description: Medicine name
        dosage:
          type: string
          description: Medicine dosage
        frequency:
          type: string
          description: How often to take
        stock:
          type: integer
          description: Number of pills in stock
        low_stock_threshold:
          type: integer
          description: Alert threshold
        time_windows:
          type: array
          items:
            type: object
            properties:
              hour:
                type: integer
                minimum: 0
                maximum: 23
              minute:
                type: integer
                minimum: 0
                maximum: 59
        active:
          type: boolean
          default: true
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the response was generated
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
        meta:
          $ref: '#/components/schemas/Meta'
    PaginatedMedicineList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Medicine'
        meta:
          type: object
          properties:
            total:
              type: integer
            count:
              type: integer
            page:
              type: integer
            per_page:
              type: integer
            total_pages:
              type: integer
            timestamp:
              type: string
              format: date-time
...